sprint-4/step-1 - extract ingredients reducer
