+++ src/services/reducers/connectors/IngredientService.ts
@@ -1,0 +1,14 @@
+import { IBurgerPart } from '../../model/IBurgerPart';
+
+export class IngredientService {
+  private static ingredients: IBurgerPart[] = [];
+
+  static save(ingredients: IBurgerPart[]) {
+    IngredientService.ingredients = ingredients;
+  }
+
+  static getIngredients(): IBurgerPart[] {
+    return IngredientService.ingredients;
+  }
+
+}
