+++ src/services/reducers/ingredients/index.ts
@@ -1,0 +1,2 @@
+export { ingredientsReducer } from './ingredientsReducer';
+export type { IngredientsState } from './IngredientsState'
