+++ src/services/reducers/main/index.ts
@@ -1,0 +1,2 @@
+export { mainReducer } from './mainReducer';
+export type { IMainState } from './IMainState'
